requirements.txt
# Hansel - Product Discovery AI Agent Team
# Python 3.11+

# UI Framework
streamlit>=1.30.0

# LLM Providers
anthropic>=0.18.0
google-generativeai>=0.4.0

# Data Validation
pydantic>=2.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
python-frontmatter>=1.0.0    # Markdown frontmatter parsing
